requirements.txt
Django==4.2.23
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.0
python-decouple==3.8
Pillow==10.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
pycryptodome==3.19.0
python-dotenv==1.0.0
requests==2.32.4
celery==5.5.3
redis==6.1.1
google-generativeai
pandas
numpy
